changed clean to make it compatible with half run varMDsim

+++ snp2sim.py
@@ -1477,11 +1477,12 @@ def runVarMDsim(parameters):
 			print("ERROR - %s-%s is already created... resubmit with new protein/variant combination" %(parameters.protein, parameters.variant))
 			sys.exit()
 		if hasattr(parameters, "clean") and parameters.clean:
-			shutil.rmtree('%s/variantSimulations/%s/results/%s' %(parameters.runDIR,parameters.protein, parameters.variant))
+			if os.isdir('%s/variantSimulations/%s/results/%s' %(parameters.runDIR,parameters.protein, parameters.variant)):
+				shutil.rmtree('%s/variantSimulations/%s/results/%s' %(parameters.runDIR,parameters.protein, parameters.variant))
 			os.makedirs('%s/variantSimulations/%s/results/%s' %(parameters.runDIR,parameters.protein, parameters.variant))
 		if not os.path.isdir("%s/variantSimulations/%s/structures" % (parameters.runDIR,parameters.protein)):
 			os.makedirs("%s/variantSimulations/%s/structures" % (parameters.runDIR,parameters.protein))
-			os.system("cp %s %s/variantSimulations/%s/structures/%s.template.pdb" \
+		os.system("cp %s %s/variantSimulations/%s/structures/%s.template.pdb" \
 				  % (parameters.newStruct,parameters.runDIR,
 					 parameters.protein, parameters.protein))
